+++ CHANGELOG.md
@@ -25,6 +25,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 ### Features
 
 - shared: use `five8` to encode/decode ([#70](https://github.com/lamports-dev/richat/pull/70))
+- shared: use `slab` for `Shutdown` ([#75](https://github.com/lamports-dev/richat/pull/75))
 
 ### Breaking